Исправления предупреждений - double to float и переопределения

+++ xrEngine/x_ray.cpp
@@ -1353,8 +1353,8 @@ u32 calc_progress_color(u32 idx, u32 total, int stage, int max_stage)
 	if (idx > (total / 2))
 		idx = total - idx;
 
-	double kk = (float(stage + 1) / double(max_stage)) * (total / 2.0f);
-	double f = 1 / (exp((double(idx) - kk) * 0.5f) + 1.0f);
+	float kk = (float(stage + 1) / float(max_stage)) * (total / 2.0f);
+	float f = 1 / (expf((float(idx) - kk) * 0.5f) + 1.0f);
 
 	return color_argb_f(f, 1.0f, 1.0f, 1.0f);
 }
